feat: Add DrawingSystems and improve documentation
