+++ ryot/src/bevy_ryot/drawing/commands/update.rs
@@ -1,62 +1,18 @@
-use crate::bevy_ryot::drawing::{
-    CommandState, CommandType, Deletion, DrawingBundle, ReversibleCommand,
-};
-use crate::bevy_ryot::map::MapTiles;
-use crate::bevy_ryot::AppearanceDescriptor;
-use crate::prelude::drawing::TileComponent;
-use crate::prelude::TilePosition;
-use crate::Layer;
+use crate::bevy_ryot::drawing::*;
 use bevy::ecs::system::Command;
-use bevy::prelude::*;
 
-#[cfg(feature = "lmdb")]
-use crate::bevy_ryot::lmdb::LmdbEnv;
-#[cfg(feature = "lmdb")]
-use crate::lmdb::{GetKey, Item, ItemRepository, ItemsFromHeedLmdb, Tile};
-#[cfg(feature = "lmdb")]
-use bevy::log::error;
-#[cfg(feature = "lmdb")]
-use std::collections::HashMap;
-
-pub type DrawingInfo = (
-    TilePosition,
-    Layer,
-    Visibility,
-    Option<AppearanceDescriptor>,
-);
-
-impl From<DrawingBundle> for DrawingInfo {
-    fn from(bundle: DrawingBundle) -> Self {
-        (
-            bundle.tile_pos,
-            bundle.layer,
-            bundle.visibility,
-            Some(bundle.appearance),
-        )
-    }
-}
-
-#[derive(Eq, PartialEq, Component, Default, Copy, Clone)]
-pub struct UpdateComponent {
-    pub new: DrawingInfo,
-    pub old: DrawingInfo,
-    pub state: CommandState,
-}
-
-impl UpdateComponent {
-    pub fn new(new: DrawingInfo, old: DrawingInfo) -> Self {
-        Self {
-            new,
-            old,
-            state: CommandState::default(),
-        }
-    }
-}
-
+/// The main command for manipulating the content of a tile. It is used to update the content of a
+/// tile, by adding, removing or updating the entities that are drawn on the tile. It is also used
+/// to undo and redo the changes made to the tile content.
+///
+/// The command holds two vectors of DrawingInfo, one for the new content and another for the old
+/// content. The new content is the content that will be applied to the tile, while the old content
+/// is the content that will be reverted to, when/if the command is undone.
 #[derive(Debug, Clone)]
 pub struct UpdateTileContent(pub Vec<DrawingInfo>, Vec<DrawingInfo>);
 
 impl UpdateTileContent {
+    /// Constructor that guarantees that the new and old content have the same length.
     pub fn new(new: Vec<DrawingInfo>, old: Vec<DrawingInfo>) -> Self {
         if new.len() != old.len() {
             panic!("The new and old content must have the same length");
@@ -65,6 +21,7 @@ impl UpdateTileContent {
         Self(new, old)
     }
 
+    /// Constructor that inits the command from a clean state, when only the new content is known.
     pub fn for_new_bundle(bundles: Vec<DrawingBundle>) -> Self {
         Self::new(
             bundles
@@ -79,6 +36,7 @@ impl UpdateTileContent {
         )
     }
 
+    /// Reverts the command, by swapping the new and old content. Useful for undoing the command.
     pub fn revert(&self) -> Self {
         Self::new(self.1.clone(), self.0.clone())
     }
@@ -99,28 +57,8 @@ impl Command for UpdateTileContent {
                 continue;
             }
 
-            let id = get_or_create_entity_for_info(world, info);
-
-            world
-                .entity_mut(id)
-                .insert(UpdateComponent::new(*info, old[index]));
+            update(world, *info, old[index], CommandState::default());
         }
-    }
-}
-
-pub fn get_or_create_entity_for_info(world: &mut World, info: &DrawingInfo) -> Entity {
-    let (pos, layer, ..) = info;
-
-    let entity = world
-        .resource_mut::<MapTiles>()
-        .entry(*pos)
-        .or_default()
-        .get(layer)
-        .copied();
-
-    match entity {
-        Some(entity) => entity,
-        None => world.spawn_empty().id(),
     }
 }
 
@@ -131,110 +69,5 @@ impl ReversibleCommand for UpdateTileContent {
 
     fn redo(&self, commands: &mut Commands) {
         commands.add(self.clone());
-    }
-}
-
-pub fn apply_update(
-    mut commands: Commands,
-    mut tiles: ResMut<MapTiles>,
-    mut q_inserted: Query<
-        (Entity, &mut UpdateComponent),
-        Or<(Changed<UpdateComponent>, Added<UpdateComponent>)>,
-    >,
-) {
-    for (entity, mut update) in q_inserted.iter_mut() {
-        if update.state.applied {
-            continue;
-        }
-
-        let (pos, layer, visibility, appearance) = update.new;
-
-        // If no appearance is provided, update is ended and the deletion is triggered.
-        let Some(appearance) = appearance else {
-            commands
-                .entity(entity)
-                .insert(Deletion::default())
-                .remove::<UpdateComponent>();
-
-            continue;
-        };
-
-        commands
-            .entity(entity)
-            .insert((pos, layer, appearance, visibility, TileComponent))
-            .remove::<Deletion>();
-
-        tiles.entry(pos).or_default().entry(layer).or_insert(entity);
-
-        update.state.applied = true;
-    }
-}
-
-pub fn persist_update(
-    #[cfg(feature = "lmdb")] lmdb_env: Res<LmdbEnv>,
-    mut q_inserted: Query<
-        &mut UpdateComponent,
-        Or<(Changed<UpdateComponent>, Added<UpdateComponent>)>,
-    >,
-) {
-    #[cfg(feature = "lmdb")]
-    {
-        let mut keys = vec![];
-        let mut to_draw = vec![];
-
-        for update in q_inserted.iter_mut() {
-            let (tile_pos, layer, _, appearance) = update.new;
-
-            if update.state.persisted {
-                continue;
-            }
-
-            let Some(appearance) = appearance else {
-                continue;
-            };
-
-            keys.push(tile_pos.get_binary_key());
-            to_draw.push((tile_pos, layer, appearance));
-        }
-
-        let item_repository = ItemsFromHeedLmdb::new(lmdb_env.clone());
-        let mut new_tiles: HashMap<TilePosition, Tile> = HashMap::new();
-
-        let tiles = item_repository.get_for_keys(keys);
-
-        if let Err(e) = tiles {
-            error!("Failed to get tiles: {}", e);
-            return;
-        };
-
-        for tile in tiles.unwrap() {
-            new_tiles.insert(tile.position, tile);
-        }
-
-        for (tile_pos, layer, appearance) in &to_draw {
-            let tile = new_tiles
-                .entry(*tile_pos)
-                .or_insert(Tile::from_pos(*tile_pos));
-
-            tile.set_item(
-                Item {
-                    id: appearance.id as u16,
-                    attributes: vec![],
-                },
-                *layer,
-            );
-        }
-
-        if let Err(e) = item_repository.save_from_tiles(new_tiles.into_values().collect()) {
-            error!("Failed to save tile: {}", e);
-        }
-    }
-
-    for mut update in q_inserted.iter_mut() {
-        if update.state.persisted {
-            continue;
-        }
-
-        update.state.persisted = true;
     }
 }
