+++ ryot_compass/src/bevy_compass/drawing/mod.rs
@@ -105,13 +105,6 @@ impl<C: ContentAssets> Plugin for DrawingPlugin<C> {
                     (draw_on_click::<C>(), draw_on_hold::<C>()),
                     toggle_deletion.run_if(action_just_pressed(DrawingAction::ToggleDeletion)),
                     (
-                        apply_update,
-                        persist_update,
-                        apply_deletion,
-                        persist_deletion,
-                    )
-                        .chain(),
-                    (
                         tick_undo_redo_timer,
                         undo_on_hold(),
                         undo_on_click(),
